refactor: #1 change Login Req/Res class position
