+++ src/main/java/laughcandidate/yellowribbonbe/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package laughcandidate.yellowribbonbe.auth.dto;
+package laughcandidate.yellowribbonbe.auth.dto.request;
 
 public record LoginRequest(
     String id,
