+++ src/main/java/laughcandidate/yellowribbonbe/auth/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package laughcandidate.yellowribbonbe.auth.dto;
+package laughcandidate.yellowribbonbe.auth.dto.response;
 
 import lombok.Builder;
